Add: confirmation avant suppression

+++ website/resources/js/components/admin/admin-list.jsx
@@ -21,9 +21,11 @@ class AdminList extends React.Component {
     }
 
     deleteAdmin(e) {
-        const id = e.target.value
-        axios.delete(`${this.props.url}/api/admin/${id}?api_token=${this.props.api}`)
-        this.componentDidMount()
+        if(confirm('Voulez-vous vraiment supprimer cet administrateur ? Cett action est irréversible')) {
+            const id = e.target.value
+            axios.delete(`${this.props.url}/api/admin/${id}?api_token=${this.props.api}`)
+            this.componentDidMount()
+        }
     }
 
     render() {
